feat: refactor VegaCellContent to use VegaCellHeader for improved structure and maintainability
